Fix config IT for nacos2
