+++ test/src/test/java/com/alibaba/nacos/test/config/ConfigLongPoll_CITCase.java
@@ -41,7 +41,7 @@ import java.util.concurrent.TimeUnit;
  * @date 2019-06-07 22:24
  **/
 @RunWith(SpringRunner.class)
-@SpringBootTest(classes = Nacos.class, properties = {"server.servlet.context-path=/nacos", "server.port=7004"},
+@SpringBootTest(classes = Nacos.class, properties = {"server.servlet.context-path=/nacos"},
         webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT)
 public class ConfigLongPoll_CITCase {
